firefox: Fix running config.status
In the previous patch the shebang of the config.status was
replaced with host's python because the build system generates
too long shebang which isn't recognized by Linux OS. But this
solution will cause other problems because the python interpreter
is different from the one which is used in the build system.

The new solution explicitly calls the python interpreter which is
used in the build system intead of replacing shebang.

Signed-off-by: Takuro Ashie <ashie@homa.ne.jp>