recipes-rgz: Recipes for overriding RZ/G series' recipes
We need the glibc patch to use WebGL2 on Firefox 68 on RZ/G series.
It should be included in meta-rzg2 or the layer extending it, but we
have decided to include the patch in meta-browser because this patch is
small enough and we have no plan to extend other recipes.